ci: Introduced vetting test on PRs and corrected package version to match "current state of the package" (#3549)
This PR introduces vetting test that should be running on PR to confirm
that the package will be in "releasable state". Things to note are

## Bumping package version
After N4E investigation in I noticed that vetting tests were not working
properly because after each release we should bump package version to
match "current state of the package" which means bumping it by a patch.
That way the current package version is being compared against last
released. Otherwise this test won't give us any valid result because we
will be comparing 2 already released packages.

Because of that I bumped NGO internal version to 2.5.0 and in that way
we will be comparing this "current" version against last released 2.4.3

## What this test will do
This test will mostly evaluate API compatibility. What we want to
achieve is to not release any new minor version if not needed/detect all
new APIs earlier. Notice that this test will fail if any new API is
introduced (because for new API we need to release new minor and not
patch) and in such scenario we should either
- Bump package version to new minor, indicating that we will release new
minor version
- Make the API internal/private

This will ensure that we know the current state of the package version.
Notice that in theory we could also use wrench/api-validation test but
this would require manually modifying this file after every wrench
update (to add release.py execution) so easier approach will be to
create a dedicated test and leave wrench one for actual release
validation

## release.py script
This python script is basically copied from N4E repo and updates
CHANGELOG and any other places that need to be formatted for the
release. For now it's only used for this test (to avoid it throwing an
error about CHANGELOG being improperly formatted) but in the future we
can use it to automate releasing

## Backport

https://github.com/Unity-Technologies/com.unity.netcode.gameobjects/pull/3550