fix: NetworkBehaviour inspector view exceptions when using a distributed authority network topology (#3578)
This PR addresses an issue where `NetworkBehaviour` could throw an
exception if viewed within the inspector view during a distributed
authority network topology based session and using a DAHost that is not
the authority of the associated `NetworkObject`.

This PR also addresses the issue where the authority does not have
editable inspector view fields when viewed by the authority during a
distributed authority session.


## Changelog

-Fixed: Issue where viewing a `NetworkBehaviour` with one or more
`NetworkVariable` fields could throw an exception if running a
distributed authority network topology with a local (DAHost) host and
viewed on the host when the host is not the authority of the associated
`NetworkObject`.

-Fixed: Issue when using a distributed authority network topology and
viewing a `NetworkBehaviour` with one or more `NetworkVariable` fields
in the inspector view would not show editable fields.

## Testing and Documentation

- No tests have been added.
- No documentation changes or additions were necessary.
- Manually testing this is required to validate.

<!-- Uncomment and mark items off with a * if this PR deprecates any
API:
### Deprecated API
- [ ] An `[Obsolete]` attribute was added along with a `(RemovedAfter
yyyy-mm-dd)` entry.
- [ ] An [api updater] was added.
- [ ] Deprecation of the API is explained in the CHANGELOG.
- [ ] The users can understand why this API was removed and what they
should use instead.
-->

## Backport

This is a v2.x distributed authority specific issue. No back port is
needed.

<!-- If this is a backport:
 - Add the following to the PR title: "\[Backport\] ..." .
 - Link to the original PR.
If this needs a backport - state this here
If a backport is not needed please provide the reason why.
If the "Backports" section is not present it will lead to a CI test
failure.
-->