chore: PR template and validation updates (#3582)
## Purpose of this PR
This PR aims to increase the quality of our PRs by enforcing certain
considerations and setting examples via PR template. This will also help
with Playtesting and validation of our changes by providing a better way
to evaluate PRs for potential playtest specification or release
blocking.
It is similar to an "official" suggestion from
https://github.com/orgs/community/discussions/84771

New PR template will be working as a "guide" on what would be nice to
have in it but the content in the template is more of an example. What
matters are sections:

1. **Documentation** --> Should describe what documentation
changes/corrections followed the code changes. This aims to help
catching any missing docs but if no documentation changes (API docs,
Documentation~) were needed then you can just acknowledge this fact in
this section
2. **Testing and QA** --> Should explain what testing coverage was added
for the code changes. This helps to understand if a given change will
require playtesting or is it fully covered by automated tests. This can
range from "a new sample was created with..." to "edge case issue that
is covered by added automated test". Note that **WE SHOULD ALMOST
ALWAYS** add some form of testing. The proposed content of this section
helps to see if you performed some manual testing (if you created some
project it's always worth linking but no need), it also helps with with
general evaluation of changes.
3. **Backport** --> A section that we already have. Should enforce
consideration if those changes should be ported to
`develop`/`develop-2.0.0` branches respectively
4. **Jira ticket** section was added for consideration if such PR is big
enough to require Jira ticket present. Note that this section is not
mandatory but may be flagged by QA when evaluating PRs for release. I
would assume that if PR takes more than a day of work then it should
have a corresponding Jira ticket

Those sections make it WAY easier to evaluate PRs that landed when we
want to release the package. It allows to see developer thinking on
important areas and works as a reminder for them so in case of looking
at the PRs it's easier for the QA to analyze those.
Important to note are :

- PR template was updated with more sections where **Documentation**,
**Testing and QA**, and **Backport** sections are **REQUIRED** for a PR
to be merged.
- I renamed the git workflow and I will update the name in the settings
when merging this PR so for now you can see backport-verification
pending since it doesn't exist anymore.
- **Testing and Documentation** section was basically divided into
separate **Documentation** and **Testing and QA sections**

Such updated PR template will look as follows
<img width="1388" height="1855" alt="image"
src="https://github.com/user-attachments/assets/7fe2f129-0440-48c3-98a3-b0eb9bae8e62"
/>

### Jira ticket
MTT-12822

### Documentation
No documentation changes or additions were necessary.

## Testing & QA
This change won't affect users and focuses around our workflow and
templates so no testing is needed

## Backport

https://github.com/Unity-Technologies/com.unity.netcode.gameobjects/pull/3583