update
Renamed the new method `ToggleParameterSync` to `EnableParameterSynchronization`. Added  overridden version of `EnableParameterSynchronization` that accepts a string for the name and changed the original method to accept the hash value of the parameter name.

Did some minor clean up in the `NetworkAnimatorStateChangeHandler.NetworkUpdate` method.

Added some additional comments in various places.