update
Reverted the change in HiddenScriptEditor.
Moved away from a NetworkAnimator custom editor to a custom property drawer to handle displaying the Animator parameters to be synchronized.
Handling the population of parameters and populating the NetworkAnimator.AnimatorParameterEntries during OnValidate.
Had to re-apply unchecking the excluded parameter in some prefabs for testing purposes.