chore: Add documentation for NetworkPrefab InstantiateWithData (#3574)
Documentation for #3497 

## Changelog

- Added: Documentation for the Network prefab handler.
- Fixed: Updated the Object spawning documentation page to be DA
compatible.

## Documentation

- Includes documentation for previously-undocumented public API entry
points.
- Includes edits to existing public API documentation.

## Testing & QA
[//]: #  (
This section is REQUIRED and should describe how the changes were tested
and how should they be tested when Playtesting for the release.
It can range from "edge case covered by unit tests" to "manual testing
required and new sample was added".
Expectation is that PR creator does some manual testing and provides a
summary of it here.)

### Functional Testing
[//]: # (If checked, List manual tests that have been performed.)
_Manual testing :_
- [ ] `Manual testing done`

_Automated tests:_
- [ ] `Covered by existing automated tests`
- [ ] `Covered by new automated tests`

_Does the change require QA team to:_

- [ ] `Review automated tests`?
- [ ] `Execute manual tests`?

If any boxes above are checked, please add QA as a PR reviewer.

## Backport

This is 2.x only documentation update and doesn't need a backport.

---------

Co-authored-by: Noel Stephens <noel.stephens@unity3d.com>
Co-authored-by: Amy Reeve <amy.reeve@unity3d.com>
Co-authored-by: Michał Chrobot <124174716+michalChrobot@users.noreply.github.com>