ci: Wrench update to 1.0.0 (#3606)
## Purpose of this PR
This PR aims to update Wrench to 1.0.0 which fixes some issues with
latest editor in Wrench scripts/Packageworks. We need to bump this
before branching to new release to mitigate errors that we will
encounter otherwise

After updating the version I needed to bump several other
PackageReferences in order for it to be compatible and I regenerated
recipes. For this I also noticed that BuilderScripts need to be out of
CI folder so I moved them and corrected paths

### Jira ticket
N/A

### Changelog
N/A

## Documentation
- No documentation changes or additions were necessary.

## Testing & QA
Green CI run is enough

## Backport

https://github.com/Unity-Technologies/com.unity.netcode.gameobjects/pull/3607