ci: Update of iOS model to 13 (#3611)
This PR aims to address iOS test failures on trunk that are happening
because iOS min version was bumped to 15.0 where **SE** model that we
were using is supporting 13.6.1 as per documentation.

To address this we will update the model to **13** which supports 16.6
so minimum spec will be satisfied and we will use newer device

Additionally this requires to address an issue of:

> Automatic UI interruption handling is available for iPhones running
iOS 15 and above (models SE-Gen3 and 13). It is enabled by default when
using those devices. Otherwise, system alerts (e.g. “Local Network
Access” permission alert, introduced in iOS 14) might cause disruptions
during test execution.
>
> If building of the test app is done on a separate (“Build”) job,
please make sure that that job has environment variable
UNITY_HANDLEUIINTERRUPTIONS set to 1.

N/A

Added mention in CI comments

Green CI for iOS test for trunk will be enough

Will do