fix: Arm64 test failures (#3619)
## Purpose of this PR
Fixes some test failures when running against arm64:
- Increased the timeout duration because the raspberry pi that tests
were run on is quite slow and sometimes doesn't make the timeout
- Removed booleans from structures that are created via random bytes
because Arm64 doesn't handle booleans the same way as x86, and bool(1)
doesn't compare as equal to bool(2) like it does on x86
- Removed booleans from the network variable HashSet test because it
created an infinite loop, because bool(1) and bool(2) DO apparently HASH
to the same key on arm64, where they create different keys on x86. (This
was never a valid test to begin with, because creating a hash set with
32 boolean key values is, strictly speaking, not possible, and only
worked by accident on x86)
- 

## Documentation
[//]: # (
This section is REQUIRED and should mention what documentation changes
were following the changes in this PR.
We should always evaluate if the changes in this PR require any
documentation changes.
)

- No documentation changes or additions were necessary.

## Testing & QA
[//]: #  (
This section is REQUIRED and should describe how the changes were tested
and how should they be tested when Playtesting for the release.
It can range from "edge case covered by unit tests" to "manual testing
required and new sample was added".
Expectation is that PR creator does some manual testing and provides a
summary of it here.)

### Functional Testing
[//]: # (If checked, List manual tests that have been performed.)
_Manual testing :_
- [ ] `Manual testing done`

_Automated tests:_
- [x] `Covered by existing automated tests`
- [ ] `Covered by new automated tests`

_Does the change require QA team to:_

- [ ] `Review automated tests`?
- [ ] `Execute manual tests`?

If any boxes above are checked, please add QA as a PR reviewer.

## Backport
[//]: # (
This section is REQUIRED and should link to the PR that targets other
NGO version which is either develop or develop-2.0.0 branch
Add the following to the PR title: "\[Backport\] ..."
If this is not needed, for example feature specific to NGOv2.X, then
just mention this fact.
)
Not needed.

Co-authored-by: Noel Stephens <noel.stephens@unity3d.com>