ci: Improvement of PR template (#3628)
## Purpose of this PR
Following on N4E investigation I added a functionality of automatic
assignement of netcode-qa as a PR reviewer if one of the checkboxes is
marked.

Additionally note that **Jira Ticket** section is now mandatory buy of
course you can always write N/A under it

I also modified github action runners to use the recommended one

### Jira ticket
MTT-12822

## Documentation
Included comments in the scripts

## Testing & QA
Will check one of the checkboxes to see if netcode-qa is being properly
added as reviewer, other then that no specific testing is required

### Functional Testing

_Manual testing :_
- [ ] `Manual testing done`

_Automated tests:_
- [ ] `Covered by existing automated tests`
- [ ] `Covered by new automated tests`

_Does the change require QA team to:_

- [ ] `Review automated tests`?
- [ ] `Execute manual tests`?
- [x] `Provide feedback about the PR`

If any boxes above are checked the QA team will be automatically added
as a PR reviewer.

## Backports

https://github.com/Unity-Technologies/com.unity.netcode.gameobjects/pull/3629

---------

Co-authored-by: Emma <emma.mcmillan@unity3d.com>