ci: Setting up conditional CI runs depending on PR comment trigger (#3580)
This PR aims to address the issue of "overdoing" the CI tests which due
to distribution times often lead to very long wait times.

First thing is that I failed to notice that `targets` and `drafts`
conditions are configured as separate items in a list (see PR job
trigger changes). In our CI, basic triggers in a list are combined with
an OR operator, meaning the job will run if any of the conditions are
met. This lead to test being executed on draft PRs.

What I'm changing is

1. Tests will correctly **not run** on draft PRs and if PR is not
targeting one of specified branches
2. PR trigger tests will execute only if there are changes in package or
project. For for example DOCS only changes no tests will run (in the
future we can improve that if for example we would have a test that
verifies typos etc)

This change should improve our workflow with CI as well as save some
resources for others

**NOTE THAT**
By default the PR trigger it's triggered if

1. PR targets develop or release branches
2. PR is not a draft
3. PR changes files in package or TestProjects folders (doesn't run on
for example DOCS only changes)

Note that in other cases you can trigger it by writing a comment `/ci
ngo` in the PR thread (if for example you want to test it on draft PR)

## Additional stuff
I also used that occasion to remove disable-burst file. We are not using
it and I included more advanced version in
https://github.com/Unity-Technologies/com.unity.netcode.gameobjects/pull/3557
so if we ever need it we can reuse that one. Because of that I will
remove it to avoid any confusion

## Steps after approval
I will remove the mandatory PR trigger job from branch settings since
now it will not always run (eg. only DOCS changes) and would block PRs
otherwise

## Backports

https://github.com/Unity-Technologies/com.unity.netcode.gameobjects/pull/3581

## Testing & QA
I verified that

1. PR trigger gets executed correctly when we have package or project
changes
4. I verified that the fact of tests running blocks PR from merging

## Documentation
I added explanation in jobs comments

## Jira ticket
MTT-12903