fix: [Backport] Add help urls (#3676)
## Purpose of this PR
Adds help urls to all components that are added to the component menu.
This ensures the question mark icon in the inspector view will redirect
to a valid page.

### Jira ticket
[MTTB-1452](https://jira.unity3d.com/browse/MTTB-1452)

### Changelog

- Added: Help URLs ensure the inspector redirects to valid web pages.

## Documentation

- No documentation changes or additions were necessary.

## Testing & QA (How your changes can be verified during release
Playtest)
[//]: #  (
This section is REQUIRED and should describe how the changes were tested
and how should they be tested when Playtesting for the release.
It can range from "edge case covered by unit tests" to "manual testing
required and new sample was added".
Expectation is that PR creator does some manual testing and provides a
summary of it here.)

<!-- Add any performance testing results here if relevant. -->

### Functional Testing
[//]: # (If checked, List manual tests that have been performed.)
_Manual testing :_
- [x] `Manual testing done`

_Automated tests:_
- [ ] `Covered by existing automated tests`
- [x] `Covered by new automated tests`

_Does the change require QA team to:_

- [ ] `Review automated tests`?
- [ ] `Execute manual tests`?
- [ ] `Provide feedback about the PR`?

If any boxes above are checked the QA team will be automatically added
as a PR reviewer.

## Backports

This is a backport of #3663