fix
Only invoke OnDestroy if Singleton is not null during the application quit.
Resolved the issue where a server spawning an object with ownership that then changes the parent in the same frame with a NetworkTransform that has SwitchTransformSpaceWhenParented enabled was not applying the parenting directive (being sent to all observers) locally.