chore: Optimize NetworkObject access from NetworkBehaviour (#3687)
## Purpose of this PR

Methods inside of properties are much less performant to access that
fields or methods without properties.

This PR focusses on:
1. Removing references to `NetworkBehaviour.NetworkObject` from within
`NetworkBehaviour`. Instead using `m_NetworkObject` anywhere that will
be called after the `NetworkBehaviour` is spawned.
a. Ensure that the `NetworkObject` explicitly sets `m_NetworkManager`
when it calculates whether or not a `NetworkBehaviour` belongs in it's
`ChildNetworkBehaviours` list.
2. Simplifies the areas where the `NetworkBehaviour` lifecycle callbacks
are invoked
a. Simplifies the callsites around `NetworkBehaviour.OnLostOwnership`
and `NetworkBehaviour.OnGainedOwnership`
    b. Reduces loop duplication around `NetworkBehaviour.OnNetworkSpawn`

### Jira ticket
_Link to related jira ticket ([Use the smart
commits](https://support.atlassian.com/bitbucket-cloud/docs/use-smart-commits/)).
Short version (e.g. MTT-123) also works and gets auto-linked_
to be created.

### Changelog

- Fixed: Improved `NetworkBehaviour` performance.

## Documentation

- No documentation changes or additions were necessary.

## Testing & QA (How your changes can be verified during release
Playtest)

This change needs a full play of astroids, and potentially the
OwnershipChanging demo. Everything here should be covered by the
automated tests, but this is a hot enough codepath that playtesting is
definitely needed here.

### Functional Testing
[//]: # (If checked, List manual tests that have been performed.)
_Manual testing :_
- [x] `Manual testing done`

_Automated tests:_
- [x] `Covered by existing automated tests`
- [ ] `Covered by new automated tests`

_Does the change require QA team to:_

- [ ] `Review automated tests`?
- [ ] `Execute manual tests`?
- [ ] `Provide feedback about the PR`?

If any boxes above are checked the QA team will be automatically added
as a PR reviewer.

## Backports

This is a pure performance improvement and so doesn't need a backport.

---------

Co-authored-by: Noel Stephens <noel.stephens@unity3d.com>